Slow UI during ShipStation and CSV imports: GROOV-914

+++ app/models/order_import_summary.rb
@@ -19,13 +19,15 @@ class OrderImportSummary < ActiveRecord::Base
     result['import_info'] = self.reload
     result['import_items'] = []
     # import_items = ImportItem.where('order_import_summary_id = '+self.id.to_s+' OR order_import_summary_id is null')
-    import_items = ImportItem.all
-    import_items.each do |import_item|
-      if import_item.store.nil?
-        import_item.destroy
-      else
-        result['import_items'].push({store_info: import_item.store, import_info: import_item,
-                                     show_update: show_update(import_item.store)})
+    if self.display_summary
+      import_items = ImportItem.all
+      import_items.each do |import_item|
+        if import_item.store.nil?
+          import_item.destroy
+        else
+          result['import_items'].push({store_info: import_item.store, import_info: import_item,
+                                       show_update: show_update(import_item.store)})
+        end
       end
     end
     GroovRealtime::emit('import_status_update', result, :tenant)
